feat: add an illustrated header to the event archive page
This would be ideally user configurable but unfortunately WordPress
doesn't allow us to set an existing page as an archive page. This works
only for the main "posts page" which is apparently an exception. This
way it is at least more dynamic as we can show different headlines
depending on whether the user is browsing future or past events.