feat: move away from must-use plugins in git to manual plugin management
The WordPress Core already has to be installed & updated manually,
so from now, we'll be using the same approach for plugins as well.

The installation instructions have been updated accordingly and we have
removed WordPress Importer from must-use plugins (it is optional
anyway).

The alternative is to use composer for all dependencies, including WP
Core and the plugins. However, that'd mean that any update would require
touching the git repo which would likely add unnecessary long-term
overhead.