chore: refactor the base Latte block layout structure
A new "page-header" block was added so that it is possible to modify the
page header from a sub-template without re-implementing the layout.

This was considered before but wasn't deemed that useful as our layout
is rather simple. But as it turns out, we can save up quite a bit of
boilerplate.