chore: optimize all PNG resources
All PNG resources were optimized using pngquant which utilizes the
quantization algorithm and additional techniques to reduce the image
size. The main trick is to reduce the number of colors used in the
image.

The actual command used to optimize the images:

  $ pngquant --speed 1 --quality=50-60 --ext=".png" --force --verbose -- *.png

Higher quality was used for certain images where the high compression
became visible.