feat: self-host the site fonts (get rid of Google Fonts)
Roboto and Amatic SC are now hosted directly on the site. The impact on
performance is going to be negligible or even possibly negative as the
page won't be able to take advantage of the font being already available
in the browser cache. However, through this step the page becomes fully
self-contained with zero dependencies on external resources.